Moved logging filters into separate file
