+++ src/reserv/__init__.py
@@ -4,6 +4,8 @@ import logging
 import logging.config
 import os
 
+from reserv.tools.log_filters import ConsoleFilter, WebRequestFilter
+
 def create_app():
     app = Flask(__name__, instance_relative_config=True)
 
@@ -89,27 +91,3 @@ def create_app():
     init_app(app)
     
     return app
-
-class ConsoleFilter(logging.Filter):
-    """
-    Defines a logging filter for the development server logs
-    """
-    def filter(self, record):
-        filter_strs = [
-             "Press CTRL+C to quit",
-             "This is a development server."
-        ]
-        return not any(filter_str in record.getMessage() 
-                       for filter_str in filter_strs)
-    
-
-class WebRequestFilter(logging.Filter):
-    """
-    Defines a logging filter for all web requests made by werkzeug
-    """
-    def filter(self, record):
-        filter_strs = [
-             "HTTP/1.1"
-        ]
-        return not any(filter_str in record.getMessage() 
-                       for filter_str in filter_strs)
